created matchState, moved it there
